clases/2021-01/25_Contenedores/app_2/requirements.txt
scikit-learn
fastapi
uvicorn[standard]
mysql-connector-python